# project-docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Actor Platform API",
    "version": "0.0.1",
    "description": "API for managing actors, runs, schedules, storage, and webhooks on the actor automation platform."
  },
  "servers": [
    { "url": "http://localhost:3000", "description": "Local development" }
  ],
  "paths": {
...
    "/workspaces/{workspaceId}/actors": {
      "get": {
        "summary": "List actors",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "limit", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 10 } },
          { "name": "cursor", "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Paginated actor list", "content": { "application/json": { "schema": { "type": "object", "properties": { "actors": { "type": "array", "items": { "$ref": "#/components/schemas/Actor" } }, "nextCursor": { "type": "string" } } } } } } }
      },
      "post": {
        "summary": "Create an actor",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:write"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateActorInput" } } } },
        "responses": { "201": { "description": "Actor created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Actor" } } } } }
      }
    },
    "/workspaces/{workspaceId}/actors/{actorId}": {
      "get": {
        "summary": "Get actor by ID",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "actorId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Actor details", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Actor" } } } } }
      },
      "patch": {
        "summary": "Update an actor",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "actorId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateActorInput" } } } },
        "responses": { "200": { "description": "Actor updated", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Actor" } } } } }
      },
      "delete": {
        "summary": "Delete an actor",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "actorId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Actor deleted", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
    "/workspaces/{workspaceId}/actors/{actorId}/versions": {
      "get": {
        "summary": "List actor versions",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:read"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "actorId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Version list", "content": { "application/json": { "schema": { "type": "object", "properties": { "versions": { "type": "array", "items": { "$ref": "#/components/schemas/ActorVersion" } } } } } } } }
      }
    },
    "/workspaces/{workspaceId}/actors/{actorId}/transition": {
      "post": {
        "summary": "Transition actor lifecycle state",
        "tags": ["Actors"],
        "security": [{ "bearerAuth": ["actors:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "actorId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TransitionInput" } } } },
        "responses": { "200": { "description": "Transition applied", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Actor" } } } } }
      }
    },
    "/workspaces/{workspaceId}/api-tokens": {
      "get": {
        "summary": "List API tokens",
        "tags": ["API Tokens"],
        "security": [{ "bearerAuth": ["tokens:read"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "responses": { "200": { "description": "Token list", "content": { "application/json": { "schema": { "type": "object", "properties": { "tokens": { "type": "array", "items": { "$ref": "#/components/schemas/ApiToken" } } } } } } } }
      },
      "post": {
        "summary": "Create an API token",
        "tags": ["API Tokens"],
        "security": [{ "bearerAuth": ["tokens:write"] }],
        "parameters": [{ "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } }],
        "requestBody": { "required": true, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTokenInput" } } } },
        "responses": { "201": { "description": "Token created", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/CreateTokenResponse" } } } } }
      }
    },
    "/workspaces/{workspaceId}/api-tokens/{tokenId}/revoke": {
      "post": {
        "summary": "Revoke an API token",
        "tags": ["API Tokens"],
        "security": [{ "bearerAuth": ["tokens:write"] }],
        "parameters": [
          { "name": "workspaceId", "in": "path", "required": true, "schema": { "type": "string" } },
          { "name": "tokenId", "in": "path", "required": true, "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "Token revoked", "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SuccessResponse" } } } } }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "SHA-256 hashed API token"
      }
    },
    "schemas": {
...
      "SuccessResponse": { "type": "object", "properties": { "success": { "type": "boolean" } }, "required": ["success"] },
...
      "Actor": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "workspaceId": { "type": "string" }, "ownerId": { "type": "string" },
          "name": { "type": "string" }, "slug": { "type": "string" }, "description": { "type": "string" },
          "tags": { "type": "array", "items": { "type": "string" } }, "icon": { "type": "string" },
          "status": { "type": "string", "enum": ["DRAFT", "PUBLISHED", "DEPRECATED"] },
          "inputSchema": {}, "createdAt": { "type": "string", "format": "date-time" }, "updatedAt": { "type": "string", "format": "date-time" }
        }
      },
      "ActorVersion": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "actorId": { "type": "string" }, "version": { "type": "integer" },
          "inputSchema": {}, "sourceReference": { "type": "string" }, "changelog": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "ApiToken": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "label": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }, "lastUsedAt": { "type": "string", "format": "date-time" },
          "revokedAt": { "type": "string", "format": "date-time" }
        }
      },
...
      "CreateActorInput": {
        "type": "object",
        "required": ["name", "slug"],
        "properties": {
          "name": { "type": "string" }, "slug": { "type": "string" },
          "description": { "type": "string" }, "tags": { "type": "array", "items": { "type": "string" } },
          "icon": { "type": "string" }, "inputSchema": { "type": "object" }
        }
      },
      "UpdateActorInput": {
        "type": "object",
        "properties": {
          "name": { "type": "string" }, "slug": { "type": "string" },
          "description": { "type": "string" }, "tags": { "type": "array", "items": { "type": "string" } },
          "icon": { "type": "string" }, "inputSchema": { "type": "object" }
        }
      },
      "TransitionInput": {
        "type": "object",
        "required": ["action"],
        "properties": {
          "action": { "type": "string", "enum": ["publish", "deprecate", "republish"] },
          "changelog": { "type": "string" }
        }
      },
...
      "CreateTokenInput": {
        "type": "object",
        "required": ["label", "userId"],
        "properties": {
          "label": { "type": "string" }, "userId": { "type": "string" },
          "scopes": { "type": "array", "items": { "type": "string" } }
        }
      },
      "CreateTokenResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" }, "label": { "type": "string" }, "scopes": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" }, "token": { "type": "string" }
        }
      },
...
    }
  }
}